[merc] Add merc extraction (#1356)
* docs for ee merc code

* wip

* more extraction stuff

* partial mat1 working

* mat1

* cleanup

* partial mat2 and mat3 support

* merc extraction seems to work

+++ game/system/Deci2Server.cpp
@@ -93,7 +93,7 @@ void Deci2Server::read_data() {
   int got = 0;
 
   while (got < desired_size) {
-    ASSERT(got + desired_size < buffer.size());
+    ASSERT(got + desired_size < (int)buffer.size());
     auto x = read_from_socket(accepted_socket, buffer.data() + got, desired_size - got);
     if (want_exit_callback()) {
       return;
